+++ src/commands/hardis/work/save.ts
@@ -143,6 +143,12 @@ autoRemoveUserPermissions:
     this.gitUrl = await git().listRemote(["--get-url"]);
     this.currentBranch = await getCurrentGitBranch();
     if (this.targetBranch == null) {
+      const userConfig = await getConfig("user");
+      if (userConfig?.localStorageBranchTargets[localBranch]) {
+        this.targetBranch = userConfig?.localStorageBranchTargets[localBranch];
+      }
+    }
+    if (this.targetBranch == null) {
       this.targetBranch = await selectTargetBranch({ message: "Please select the target branch of your Merge Request" });
     }
     // User log info